+++ src/WIFI/WIFI_Controller.py
@@ -1,4 +1,5 @@
 import subprocess
+from typing import Optional
 
 import src.SystemDependency as sys_dep
 
@@ -9,39 +10,60 @@ import src.SystemDependency as sys_dep
 OPERATING_SYSTEM = sys_dep.get_os()
 
 # Check for OS
-if OPERATING_SYSTEM == "darwin":
-    # MAC OS
-    print("\nOS:\t", "Mac OS\n")
-
-elif OPERATING_SYSTEM == "win64":
-    # Windows 64-bit
-    print("\nOS:\t", "Windows\n")
+print(f"\nOS:\t {OPERATING_SYSTEM}\n")
 
 
 def get_current_ssid() -> str:
     # get SSID from Terminal
-    ssid_command = "/System/Library/PrivateFrameworks/Apple80211.framework/Resources/airport" \
-                   " -I | awk -F: '/ SSID/{print $2}'"
-    ssid = subprocess.check_output(ssid_command, shell=True)
+    ssid_command = ""
+    ssid = ""
 
-    # Clean up SSID
-    ssid = str(ssid).replace("b' ", "")
-    ssid = ssid.replace("\\n'", "")
+    if OPERATING_SYSTEM == "Mac OS":
+        ssid_command = "/System/Library/PrivateFrameworks/Apple80211.framework/Resources/airport" \
+                       " -I | awk -F: '/ SSID/{print $2}'"
+        ssid = subprocess.check_output(ssid_command, shell=True)
+
+        # Clean up SSID
+        ssid = str(ssid).replace("b' ", "")
+        ssid = ssid.replace("\\n'", "")
+
+    elif OPERATING_SYSTEM == "Windows":
+        ssid_command = "Netsh WLAN show interfaces"
+        temp_ssid = subprocess.check_output(ssid_command, shell=True)
+        ssid_information: list = str(temp_ssid).split(":")
+
+        ssid = str(ssid_information[12]).split("\\r\\n")[0]
+        ssid = ssid.strip()
+
     print(f"SSID: {ssid}")
 
     return ssid
 
 
 def get_current_password(ssid) -> str:
-    # get password for the current SSID from Terminal
-    password_command = f"security find-generic-password -wa {ssid}"
-    password = subprocess.check_output(password_command, shell=True)
+    password = ""
 
-    # Clean up password
-    password = str(password).replace("b'", "")
-    password = password.replace("\\n'", "")
+    print(f"operating system: {OPERATING_SYSTEM}")
+    if OPERATING_SYSTEM == "Mac OS":
+        # get password for the current SSID from Terminal
+        password_command = f"security find-generic-password -wa {ssid}"
+        password = subprocess.check_output(password_command, shell=True)
+
+        # Clean up password
+        password = str(password).replace("b'", "")
+        password = password.replace("\\n'", "")
+        print(f"Password: {password}")
+
+    elif OPERATING_SYSTEM == "Windows":
+        # get password for the current SSID from CMD
+        password_command = f"Netsh wlan show profile name={ssid} key=clear"
+        password = subprocess.check_output(password_command, shell=True)
+
+        # Clean up password
+        password = str(password).split(":")[21].split("\\r\\n")[0]
+        password = password.strip()
+
     print(f"Password: {password}")
-
     return password
 
 
@@ -58,15 +80,11 @@ def create_wifi_text(ssid: any, password: any) -> str:
 
 
 def try_current_wifi() -> tuple[str, str, str]:
-    try:
-        ssid = get_current_ssid()
-        password = get_current_password(ssid)
+    ssid = get_current_ssid()
 
-        wifi_text = create_wifi_text(ssid, password)
-        # TODO: maybe optional strings
+    password = get_current_password(ssid)
 
-    except:
-        print("No wifi connection")
+    wifi_text = create_wifi_text(ssid, password)
 
     wifi_tuple = (ssid, password, wifi_text)
     return wifi_tuple